pytubeGUI/requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
CurrencyConverter==0.14.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
entrypoints==0.3
filelock==3.0.12
idna==2.9
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.6.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytube3==9.6.4
pytubeX==0.1.4
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.0.35
virtualenv-clone==0.5.4
watchdog==0.10.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
